Fix wdCreateImageFromHBITMAP().
The attempt to support seemlessly alpha channel did not work properly
because GDI produces bitmaps with garbage in those bits correspnding to
the alpha channel.

We have added new function wdCreateImageFromHBITMAPWithAlpha() which
allows to interpret alpha but caller has explicitly ask for it.

Also added an example code testing the function.