init.c: Refactor.
We need to redo the initialization because WinDrawLib interface modules
do not match 1:1 to the implementation modules, from now on.

In particular, with the D2D1 back-end, we use DWrite both for the string
API and (new not yet implemented) advanced text API.

The reason for disntinguishing string and text API is that later, we
shall likely implement the text API also for GDI+ back-end using
Uniscribe, but the simple string API does not need to load it.